requirements.txt
azure-common==1.1.28
azure-core==1.25.1
azure-identity==1.11.0b3
azure-keyvault==4.2.0
azure-keyvault-certificates==4.6.0
azure-keyvault-keys==4.7.0
azure-keyvault-secrets==4.6.0
azure-storage-blob==12.13.1
azure-storage-file-datalake==12.8.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
idna==3.4
isodate==0.6.1
msal==1.19.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.1
portalocker==2.5.1
prettytable==3.4.1
pycparser==2.21
PyJWT==2.5.0
requests==2.28.1
requests-futures==1.0.0
requests-oauthlib==1.3.1
six==1.16.0
types-cryptography==3.3.23
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
